Fixed file name issues when using long genre prompts
On some file systems the file names are capped, e.g. to 255 chars. When long prompts are used, the generated file names can exceed such limits, and cause the script to fail. 

The filenames are now only using the first 150 chars from the genre prompt in order to prevent this problem.